Refactored code to be usable externally.
