+++ assets/panda3d_logo_texture/panda3d_logo_texture-base.png
